fix: improve test organization and remove redundant integration tests
- Remove indirect comment from gopkg.in/yaml.v3 dependency
- Reorganize imports in test files for better readability
- Update device names from /dev/sda1 to /dev/vda1 in swap tests
- Change build tag from integration to mock_provisioner
- Remove redundant integration test functions for swap operations
